+++ server/.env.development
@@ -1,2 +1,2 @@
-MONGODB_URI=mongodb://localhost:27017/test
+MONGODB_URI=mongodb://localhost:27017/dev
 PORT=8080
