+++ server/.env.production
@@ -1,2 +1,2 @@
+MONGODB_URI=mongodb://localhost:27017/prod
 PORT=8080
-MONGODB_URI=mongodb://10.10.10.185:27017/test
